#9 Minor changes in exceptions, and finished project
