--- src/com/company/exceptions/GradeException.java
@@ -1,8 +1,0 @@
-package com.company.exceptions;
-
-public class GradeException extends Exception {
-    public String toString(){
-        return "Please enter right number of grade";
-    }
-
-}
