--- src/com/company/exceptions/NoFacultyException.java
@@ -1,8 +1,0 @@
-package com.company.exceptions;
-
-public class NoFacultyException extends Exception {
-
-    public String toString(){
-        return "There are no faculties in university";
-    }
-}
