--- src/com/company/exceptions/NoGroupException.java
@@ -1,7 +1,0 @@
-package com.company.exceptions;
-
-public class NoGroupException extends Exception {
-    public String toString() {
-        return "There are no groups in faculty";
-    }
-}
