--- src/com/company/exceptions/NoStudentException.java
@@ -1,8 +1,0 @@
-package com.company.exceptions;
-
-public class NoStudentException extends Exception{
-    public String toString(){
-        return "There are no students in group";
-    }
-
-}
