--- src/com/company/exceptions/NoSubjectException.java
@@ -1,8 +1,0 @@
-package com.company.exceptions;
-
-public class NoSubjectException extends Exception {
-    public String toString(){
-        return "Student does not have any subject";
-    }
-
-}
