+++ src/com/university_project/Main.java
@@ -1,0 +1,13 @@
+package com.university_project;
+
+public class Main {
+    public static void main(String[] args) {
+        Service service = new Service();
+        service.universityStructure();
+        service.uniFromOneSubjectGPA();
+        service.facultyFromOneSubjectGPA();
+        service.groupFromOneSubjectGPA();
+        service.oneStudentGPA();
+    }
+
+}
