+++ src/com/university_project/exceptions/GradeException.java
@@ -1,0 +1,8 @@
+package com.university_project.exceptions;
+
+public class GradeException extends RuntimeException {
+
+    public GradeException(String message) {
+        super(message);
+    }
+}
