+++ src/com/university_project/exceptions/NoFacultyException.java
@@ -1,0 +1,7 @@
+package com.university_project.exceptions;
+
+public class NoFacultyException extends RuntimeException {
+    public NoFacultyException(String message) {
+        super(message);
+    }
+}
