+++ src/com/university_project/exceptions/NoGroupException.java
@@ -1,0 +1,7 @@
+package com.university_project.exceptions;
+
+public class NoGroupException extends RuntimeException {
+    public NoGroupException(String message) {
+        super(message);
+    }
+}
