+++ src/com/university_project/exceptions/NoStudentException.java
@@ -1,0 +1,7 @@
+package com.university_project.exceptions;
+
+public class NoStudentException extends RuntimeException{
+    public NoStudentException(String message) {
+        super(message);
+    }
+}
