+++ src/com/university_project/exceptions/NoSubjectException.java
@@ -1,0 +1,7 @@
+package com.university_project.exceptions;
+
+public class NoSubjectException extends RuntimeException {
+    public NoSubjectException(String message) {
+        super(message);
+    }
+}
